backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
praw>=7.7.0
google-api-python-client>=2.100.0
openai>=1.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
python-multipart>=0.0.6
aiofiles>=23.0.0